+++ rust/crd/src/lib.rs
@@ -152,6 +152,9 @@ const DEFAULT_MIDDLEMANAGER_GRACEFUL_SHUTDOWN_TIMEOUT: Duration =
 const DEFAULT_ROUTER_GRACEFUL_SHUTDOWN_TIMEOUT: Duration = Duration::from_minutes_unchecked(5);
 const DEFAULT_HISTORICAL_GRACEFUL_SHUTDOWN_TIMEOUT: Duration = Duration::from_minutes_unchecked(5);
 
+// Auto TLS certificate lifetime
+const DEFAULT_SECRET_LIFETIME: Duration = Duration::from_days_unchecked(7);
+
 #[derive(Snafu, Debug, EnumDiscriminants)]
 #[strum_discriminants(derive(IntoStaticStr))]
 #[allow(clippy::enum_variant_names)]
@@ -348,6 +351,7 @@ pub struct CommonRoleGroupConfig {
     pub replicas: Option<u16>,
     pub affinity: StackableAffinity,
     pub graceful_shutdown_timeout: Option<Duration>,
+    pub requested_secret_lifetime: Duration,
 }
 
 /// Container for the merged and validated role group configurations
@@ -387,6 +391,11 @@ impl MergedConfig {
                     replicas: rolegroup.replicas,
                     affinity: rolegroup.config.config.affinity.clone(),
                     graceful_shutdown_timeout: rolegroup.config.config.graceful_shutdown_timeout,
+                    requested_secret_lifetime: rolegroup
+                        .config
+                        .config
+                        .requested_secret_lifetime
+                        .unwrap_or(DEFAULT_SECRET_LIFETIME),
                 })
             }
             DruidRole::Coordinator => {
@@ -400,6 +409,11 @@ impl MergedConfig {
                     replicas: rolegroup.replicas,
                     affinity: rolegroup.config.config.affinity.clone(),
                     graceful_shutdown_timeout: rolegroup.config.config.graceful_shutdown_timeout,
+                    requested_secret_lifetime: rolegroup
+                        .config
+                        .config
+                        .requested_secret_lifetime
+                        .unwrap_or(DEFAULT_SECRET_LIFETIME),
                 })
             }
             DruidRole::Historical => {
@@ -415,6 +429,11 @@ impl MergedConfig {
                     replicas: rolegroup.replicas,
                     affinity: rolegroup.config.config.affinity.clone(),
                     graceful_shutdown_timeout: rolegroup.config.config.graceful_shutdown_timeout,
+                    requested_secret_lifetime: rolegroup
+                        .config
+                        .config
+                        .requested_secret_lifetime
+                        .unwrap_or(DEFAULT_SECRET_LIFETIME),
                 })
             }
             DruidRole::MiddleManager => {
@@ -428,6 +447,11 @@ impl MergedConfig {
                     replicas: rolegroup.replicas,
                     affinity: rolegroup.config.config.affinity.clone(),
                     graceful_shutdown_timeout: rolegroup.config.config.graceful_shutdown_timeout,
+                    requested_secret_lifetime: rolegroup
+                        .config
+                        .config
+                        .requested_secret_lifetime
+                        .unwrap_or(DEFAULT_SECRET_LIFETIME),
                 })
             }
             DruidRole::Router => {
@@ -441,6 +465,11 @@ impl MergedConfig {
                     replicas: rolegroup.replicas,
                     affinity: rolegroup.config.config.affinity.clone(),
                     graceful_shutdown_timeout: rolegroup.config.config.graceful_shutdown_timeout,
+                    requested_secret_lifetime: rolegroup
+                        .config
+                        .config
+                        .requested_secret_lifetime
+                        .unwrap_or(DEFAULT_SECRET_LIFETIME),
                 })
             }
         }
@@ -1105,6 +1134,10 @@ pub struct BrokerConfig {
     /// [graceful shutdown documentation](DOCS_BASE_URL_PLACEHOLDER/druid/usage-guide/operations/graceful-shutdown).
     #[fragment_attrs(serde(default))]
     pub graceful_shutdown_timeout: Option<Duration>,
+
+    /// Request secret (currently only auto certificates) lifetime from the secret operator.
+    #[fragment_attrs(serde(default))]
+    pub requested_secret_lifetime: Option<Duration>,
 }
 
 impl BrokerConfig {
@@ -1118,6 +1151,7 @@ impl BrokerConfig {
             logging: product_logging::spec::default_logging(),
             affinity: get_affinity(cluster_name, role, deep_storage),
             graceful_shutdown_timeout: Some(role.default_graceful_shutdown_timeout()),
+            requested_secret_lifetime: Some(DEFAULT_SECRET_LIFETIME),
         }
     }
 }
@@ -1148,6 +1182,9 @@ pub struct CoordinatorConfig {
     /// [graceful shutdown documentation](DOCS_BASE_URL_PLACEHOLDER/druid/usage-guide/operations/graceful-shutdown).
     #[fragment_attrs(serde(default))]
     pub graceful_shutdown_timeout: Option<Duration>,
+    /// Request secret (currently only auto certificates) lifetime from the secret operator.
+    #[fragment_attrs(serde(default))]
+    pub requested_secret_lifetime: Option<Duration>,
 }
 
 impl CoordinatorConfig {
@@ -1161,6 +1198,7 @@ impl CoordinatorConfig {
             logging: product_logging::spec::default_logging(),
             affinity: get_affinity(cluster_name, role, deep_storage),
             graceful_shutdown_timeout: Some(role.default_graceful_shutdown_timeout()),
+            requested_secret_lifetime: Some(DEFAULT_SECRET_LIFETIME),
         }
     }
 }
@@ -1191,6 +1229,9 @@ pub struct MiddleManagerConfig {
     /// [graceful shutdown documentation](DOCS_BASE_URL_PLACEHOLDER/druid/usage-guide/operations/graceful-shutdown).
     #[fragment_attrs(serde(default))]
     pub graceful_shutdown_timeout: Option<Duration>,
+    /// Request secret (currently only auto certificates) lifetime from the secret operator.
+    #[fragment_attrs(serde(default))]
+    pub requested_secret_lifetime: Option<Duration>,
 }
 
 impl MiddleManagerConfig {
@@ -1204,6 +1245,7 @@ impl MiddleManagerConfig {
             logging: product_logging::spec::default_logging(),
             affinity: get_affinity(cluster_name, role, deep_storage),
             graceful_shutdown_timeout: Some(role.default_graceful_shutdown_timeout()),
+            requested_secret_lifetime: Some(DEFAULT_SECRET_LIFETIME),
         }
     }
 }
@@ -1234,6 +1276,9 @@ pub struct RouterConfig {
     /// [graceful shutdown documentation](DOCS_BASE_URL_PLACEHOLDER/druid/usage-guide/operations/graceful-shutdown).
     #[fragment_attrs(serde(default))]
     pub graceful_shutdown_timeout: Option<Duration>,
+    /// Request secret (currently only auto certificates) lifetime from the secret operator.
+    #[fragment_attrs(serde(default))]
+    pub requested_secret_lifetime: Option<Duration>,
 }
 
 impl RouterConfig {
@@ -1247,6 +1292,7 @@ impl RouterConfig {
             logging: product_logging::spec::default_logging(),
             affinity: get_affinity(cluster_name, role, deep_storage),
             graceful_shutdown_timeout: Some(role.default_graceful_shutdown_timeout()),
+            requested_secret_lifetime: Some(DEFAULT_SECRET_LIFETIME),
         }
     }
 }
@@ -1277,6 +1323,9 @@ pub struct HistoricalConfig {
     /// [graceful shutdown documentation](DOCS_BASE_URL_PLACEHOLDER/druid/usage-guide/operations/graceful-shutdown).
     #[fragment_attrs(serde(default))]
     pub graceful_shutdown_timeout: Option<Duration>,
+    /// Request secret (currently only auto certificates) lifetime from the secret operator.
+    #[fragment_attrs(serde(default))]
+    pub requested_secret_lifetime: Option<Duration>,
 }
 
 impl HistoricalConfig {
@@ -1290,6 +1339,7 @@ impl HistoricalConfig {
             logging: product_logging::spec::default_logging(),
             affinity: get_affinity(cluster_name, role, deep_storage),
             graceful_shutdown_timeout: Some(role.default_graceful_shutdown_timeout()),
+            requested_secret_lifetime: Some(DEFAULT_SECRET_LIFETIME),
         }
     }
 }