+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Lifetime of auto generated certificates is configurable with the `requestedSecretLifetime` role group property ([#660])
+
 ### Fixed
 
 - Fix OIDC endpoint construction in case the `rootPath` does have a trailing slash ([#656]).
@@ -13,6 +17,7 @@ All notable changes to this project will be documented in this file.
 
 [#656]: https://github.com/stackabletech/druid-operator/pull/656
 [#657]: https://github.com/stackabletech/druid-operator/pull/657
+[#660]: https://github.com/stackabletech/druid-operator/pull/660
 
 ## [24.11.0] - 2024-11-18